refactor: track 저장 방식 변경
