+++ src/main/java/io/github/eappezo/soundary/services/music/infrastructure/persistence/JpaTrackRepository.java
@@ -2,5 +2,5 @@ package io.github.eappezo.soundary.services.music.infrastructure.persistence;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JpaTrackRepository extends JpaRepository<TrackEntity, TrackEntityKey> {
+public interface JpaTrackRepository extends JpaRepository<TrackEntity, String> {
 }
