+++ src/main/java/io/github/eappezo/soundary/services/music/infrastructure/persistence/TrackEntity.java
@@ -4,7 +4,6 @@ import io.github.eappezo.soundary.core.identification.Identifier;
 import io.github.eappezo.soundary.services.music.application.SimpleTrackDto;
 import io.github.eappezo.soundary.services.music.domain.Album;
 import io.github.eappezo.soundary.services.music.domain.Artist;
-import io.github.eappezo.soundary.services.music.domain.MusicPlatform;
 import io.github.eappezo.soundary.services.music.domain.Track;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -16,19 +15,13 @@ import java.time.Duration;
 import java.util.Arrays;
 
 @Entity(name = "tracks")
-@IdClass(TrackEntityKey.class)
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
 public class TrackEntity {
     @Id
-    @Enumerated(EnumType.STRING)
-    @Column(name = "platform")
-    private MusicPlatform platform;
-
-    @Id
-    @Column(name = "platform_track_id")
-    private String platformTrackId;
+    @Column(name = "id")
+    private String id;
 
     @Column(name = "title")
     private String title;
@@ -50,7 +43,6 @@ public class TrackEntity {
 
     public static TrackEntity from(SimpleTrackDto track) {
         return new TrackEntity(
-                track.platform(),
                 track.id(),
                 track.title(),
                 track.album(),
@@ -63,8 +55,7 @@ public class TrackEntity {
 
     public Track toDomain() {
         return Track.of(
-                platform,
-                Identifier.fromString(platformTrackId),
+                Identifier.fromString(id),
                 title,
                 Arrays.stream(artists.split(", ")).map(Artist::new).toList(),
                 new Album(albumTitle, albumCoverUrl),
